requirements.txt
py2neo
PyMySQL
tweepy
requests
python-instagram
ftfy >=4, < 5
google-api-python-client
beautifulsoup4
numpy
scipy
tinys3
nltk
beautifulsoup
